feat: add password strength for uncontrolled elements
