+++ src/components/ReactHookForm/ReactHookForm.tsx
@@ -9,12 +9,7 @@ import { save } from '../../store/dataSlice';
 import { validationSchema } from '../../utils/yup';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { useEffect, useState } from 'react';
-import {
-  lowercaseRegex,
-  numberRegex,
-  symbolRegex,
-  uppercaseRegex,
-} from '../../utils/regex';
+import { checkStrongPassword } from '../../utils/regex';
 
 interface FormInput {
   userName: string;
@@ -59,27 +54,6 @@ export function ReactHookForm() {
   const countries = useSelector(
     (state: RootState) => state.dataSlice.countries,
   );
-
-  const checkStrongPassword = (password: string): string => {
-    const hasNumber = numberRegex.test(password);
-    const hasUppercase = uppercaseRegex.test(password);
-    const hasLowercase = lowercaseRegex.test(password);
-    const hasSymbol = symbolRegex.test(password);
-    const strength = +hasNumber + +hasUppercase + +hasLowercase + +hasSymbol;
-
-    switch (strength) {
-      case 1:
-        return 'Weak password';
-      case 2:
-        return 'Average password';
-      case 3:
-        return 'Strong password';
-      case 4:
-        return 'Superb password';
-      default:
-        return '';
-    }
-  };
 
   const onSubmit: SubmitHandler<FormInput> = async (data) => {
     clearErrors();
